at_limit audit: 15 fixes to 100% from premature-AT_LIMIT triage
Six parallel fix agents audited 626 AT_LIMIT entries flagged as
potentially premature. Results: 60 re-marked COMPLETE (no code), 15
source fixes to 100% across 13 files:

F2 comparison-style (signed/unsigned compare): 9 fixes
  - Shader.cpp CheckDistortion: cache statics for cmplwi shape
  - HamCamTransform.cpp ClearOldCrowds: cache *sit local
  - ResMgr.h NumRefs/mRefs: int -> unsigned long
  - AmbientOcclusion.cpp IsSerializable: explicit (int) cast
  - LiveCameraInput.cpp Store{Color,Depth}BufferClip: u32 -> int on w/h
  - FlowMultiSetProperty.cpp Activate: switch to *it form
  - Memcard_Xbox.cpp FindValidUnit: cast mDeviceId + restructure num
    branch (target returns kMCFileNotFound when num==0, fixes latent
    behavior bug)
  - HamWardrobe.cpp PlayCrowdAnimation: (int)cd->mNext for cmpwi style
  - FlowNode.h kWhenAble: 4 -> 5 (target enum layout)

F4 super-call argument order: 2 fixes
  - CharUpperTwist.cpp Copy/Save: member order mTwist2/mUpperArm/mTwist1
    (RB3 reference confirms; was alphabetical-looking but wrong)

F6 control-flow polarity: 4 fixes
  - SkeletonChooser.cpp IsSinglePlayerMode: split && chain into if/else
  - SkeletonUpdate.cpp SkeletonUpdateThread: goto wait early-out
  - FlowSlider.cpp Load: GenerateAutoNames(this,true) replaces wrong
    bs >> mEventsRegistered read (fixes latent serialization bug)
  - SkeletonClip.cpp Poll: inline float comparison
  - Rnd.cpp TestPoint: insert at end() not begin() (95->98%, partial)

F1 commutative-op-order, F3 offset-swap, F5 sort-template clusters all
confirmed truly AT_LIMIT (regalloc, stack-spill scheduling artifacts —
not source-controllable per docs/decomp/patterns/unfixable-compiler.md).

DB delta: AT_LIMIT 4659 -> 4585 (-74), COMPLETE 27414 -> 27488 (+74).

+++ src/system/rndobj/AmbientOcclusion.cpp
@@ -464,7 +464,7 @@ bool RndAmbientOcclusion::IsSerializable(const RndMesh *mesh) const {
         return false;
     }
     ObjectDir *meshDir = mesh->Dir();
-        if ((meshDir == Dir()))
+    if ((int)meshDir == (int)Dir())
         return true;
     return (meshDir->IsSubDir() && meshDir->InlineSubDirType() == kInlineAlways);
 }
